BugzID: 3224 - CMS: Can't insert images into default Body part
On line 95 of assets.js, the part_id is being passed and ckeditor is looking for a mode. If the part_id is capitalized, for some reason this is causing a failure for ckeditor. I wasn't able to track down the specific issue, but I added a toLowerCase() on line 28 to correct this.